Move habit list related components into list/
